+++ CHANGELOG.md
@@ -4,6 +4,17 @@ This project is following [Semantic Versioning](http://semver.org)
 
 ## [Unreleased][]
 
+## [0.3.0][] - 2018-04-24
+
+### Added
+
+    - admin settings screen for configuration
+
+### Changed
+
+    - `dpapp`is now available to custom script alongside `getMe` and `getTabData` functions
+    - jquery is no longer required to load assets 
+
 ## [0.2.4][] - 2018-03-29
 
 ### Added
@@ -16,5 +27,6 @@ This project is following [Semantic Versioning](http://semver.org)
     - upgrade to @deskpro/apps-dpat version 0.10.3
 
 
-[Unreleased]: https://github.com/DeskproApps/custom-javascript/compare/v0.2.5...HEAD
+[Unreleased]: https://github.com/DeskproApps/custom-javascript/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/DeskproApps/custom-javascript/compare/v0.2.5...v0.3.0
 [0.2.4]: https://github.com/DeskproApps/custom-javascript/tree/v0.2.4